converting to single class
